adds ads. burger menu clickable without retracting. moved listenForOutsideClicks to utils.
