+++ src/assets/components/Advertisement/Ad.css
@@ -1,3 +1,7 @@
 .carousel-indicators {
     z-index: 0!important;
 }
+
+img {
+    max-height: 10rem;
+}
