--- src/assets/components/home/Home.jsx
@@ -1,15 +1,0 @@
-import RecipeRepresentation from "../RecipeRepresentation/RecipeRepresentation";
-import Search from "../Searchbar/Search";
-
-function Home() {
-  return (
-      <>
-      <Search />
-      <RecipeRepresentation />
-      </>
-      
-
-  );
-}
-
-export default Home;
